LayoutTests/crypto/workers/subtle/ec-import-pkcs8-key-expected.txt

[Worker] Test importing a PKCS8 ECDH key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/ec-import-pkcs8-key.js
[Worker] Importing a key...
PASS [Worker] publicKey.toString() is '[object CryptoKey]'
PASS [Worker] publicKey.type is 'private'
PASS [Worker] publicKey.extractable is true
PASS [Worker] publicKey.algorithm.name is 'ECDH'
PASS [Worker] publicKey.algorithm.namedCurve is 'P-384'
PASS [Worker] publicKey.usages is [ 'deriveBits', 'deriveKey' ]
PASS successfullyParsed is true

TEST COMPLETE
